% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGSAI.R
\name{RunGSAI}
\alias{RunGSAI}
\title{Generate a function to run GASI}
\usage{
RunGSAI(
  degs,
  api_key,
  gsai_prompt,
  api_url = "https://api.openai.com/v1/chat/completions",
  request_timeout_sec = 60,
  model = "gpt-4o-2024-08-06",
  temperature = 1,
  output = FALSE,
  max_retries = 1,
  min_rows = 1
)
}
\description{
Generate a function to run GASI
}
